fix: 422 float bbox + scroll drift absolute masks + dedupe
- server/app.py: bbox List[int] -> List[float] (accept 579.5 sub-pixel from getBoundingClientRect), no more 422
- content: Math.round for sensitiveInputs + text spans + image regions, dedupe 22->~14 regions via rounded key, totalRegions metric
- content: masks now position:absolute document coords (x+scrollX, y+scrollY) not fixed — survive scroll, no drift on 'Scroll down' task, scroll/resize listeners with rAF, dedupe inside render
- verify: POST with float 579.5 now 200 ok (was 422), build 22.49MB chrome + zip 5.25MB

+++ server/app.py
@@ -37,7 +37,7 @@ if PUBLIC_DIR.exists():
     def serve_ground_truth():
         return FileResponse(str(PUBLIC_DIR / "ground-truth.json"), media_type="application/json")
 
-# ---- Models ----
+# ---- Models ---- (bbox floats from getBoundingClientRect sub-pixel; accept float, coerce to int on store)
 class Viewport(BaseModel):
     width: int
     height: int
@@ -46,14 +46,14 @@ class AxNode(BaseModel):
     role: str
     name: str
     tag: str
-    bbox: List[int]
+    bbox: List[float]
     value: Optional[str] = None
     inputType: Optional[str] = None
     isSensitive: Optional[bool] = None
     placeholder: Optional[str] = None
 
 class RedactedRegion(BaseModel):
-    bbox: List[int]
+    bbox: List[float]
     type: str
     confidence: float
